exluding test cases depending on  jettyRunWar
