+++ .travis.yml
@@ -1,5 +1,5 @@
 language: java
 jdk:
 - oraclejdk7
-before_install: gradle jettyRunWar
+
