+++ service-implementation/build.gradle
@@ -184,6 +184,46 @@ jettyRun {
     webAppSourceDirectory = file('src/main/webapp/ui')
 }
 
+// jettyRun configuration properties
+jettyRunWar {
+	dependsOn 'jettyRunSetProperties'
+	logger.info "Running jetty"
+	httpPort = new Integer(jettyHttpPort)
+	daemon = true
+	stopKey='stopJetty'
+	stopPort=9999
+
+}
+
+jettyStop{
+	
+	stopKey='stopJetty'
+	stopPort=9999
+}
+
+
+
+import org.gradle.api.plugins.jetty.internal.Monitor
+[jettyRun, jettyRunWar]*.doLast {
+  /**
+   * THIS IS A WORKAROUND! THE CURRENT VERSION OF THIS TASK DOESN'T START A WATCHER IN DAEMON MODE
+   *
+   * If starting the monitor fails, it may be because the jetty task was updated to fix this issue
+   * When that happens, we shouldn't need the custom task any more
+   *
+   * Copied From: AbstractJettyRunTask
+   */
+   if (getStopPort() != null && getStopPort() > 0 && getStopKey() != null) {
+	  Monitor monitor = new Monitor(getStopPort(), getStopKey(), server.getProxiedObject());
+	  monitor.start();
+   }
+}
+
+test { // TODO temp exclusion of j2ep
+	exclude '**/FS2ResourceTest.class'
+	exclude '**/HelloWorldResourceTest.class'}
+
+
 task pingTest {
 	dependsOn 'clean'
 	dependsOn 'jettyRunWar'
