+++ service-implementation/src/main/resources/hello-world-dev.properties
@@ -1,1 +1,1 @@
-netflix.platform.admin.resources.disable=true
+netflix.platform.admin.resources.disable=false
